Texture Rendering (Still WIP)
We now have a mostly working texture pipeline with component support.
We can now also load textures with a single function quite simply.
Also some improvements to side features like transforms and the input manages which now have more supporting functions.
Currently textures don't render correctly but they at least do render.

+++ brackens_tools/src/upkeep.rs
@@ -8,8 +8,7 @@ pub struct UpkeepTracker {
     elapsed_time: Instant,
     last_frame_instant: Instant,
 
-    // If my math is correct, u32 should last for about 0.8 of a year as 150 fps :D
-    total_frame_count: u32,
+    total_frame_count: u64,
     second_tracker: f32,
     frame_count_this_second: u16,
